add authentication using jwt

+++ controllers/api/v1/posts_api.js
@@ -30,20 +30,23 @@ module.exports.destroy= async function(req,res){
     try{
         let post=await Post.findById(req.params.id);
 
-        // if (post.user==req.user.id){
+      if(post.user==req.user.id){
         post.remove()
     
         await Comment.deleteMany({post:req.params.id});
 
                
                
-        // }else{
-        //     req.flash('error','you cannot delete this post');
-        //     return res.redirect('back');
-        //     }
+        
         return res.json(200,{
             message:"Post and associated comments deleted!"
-        })
+        });
+      }else{
+          return res.json(401,{
+            message:'you cannot delete this post'
+          });
+      }
+        
         
     }catch(err){
